+++ csharp/SupermarketReceipt/FakeCatalog.cs
@@ -4,16 +4,16 @@ namespace SupermarketReceipt
 {
     public class FakeCatalog : SupermarketCatalog
     {
-        private readonly IDictionary<string, double> _prices = new Dictionary<string, double>();
+        private readonly IDictionary<string, decimal> _prices = new Dictionary<string, decimal>();
         private readonly IDictionary<string, Product> _products = new Dictionary<string, Product>();
 
-        public void AddProduct(Product product, double price)
+        public void AddProduct(Product product, decimal price)
         {
             _products.Add(product.Name, product);
             _prices.Add(product.Name, price);
         }
 
-        public double GetUnitPrice(Product p)
+        public decimal GetUnitPrice(Product p)
         {
             return _prices[p.Name];
         }
